+++ src/main/java/com/pfk/thecats/infra/thecatapi/openfeign/CatOpenFeignConfig.java
@@ -1,4 +1,4 @@
-package com.pfk.thecats.infra.thecatapi;
+package com.pfk.thecats.infra.thecatapi.openfeign;
 
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
